Added Comments to all files
Right now the issue i'm facing is reading everything comma by comma and storing it into nodes efficiently. The original idea was to read it line by line and then separate comma by comma but that takes too long. An idea is to put each comma into a node and a linked list and combine the linked list into a stock node, but i'm not sure how long that'll take. Then the next issue we need to solve is being able to tell which stocks are shown in every file. I have two ideas, one is to put everything in on linked list and count everything, which may take a long time, and the second is to create a linked list for each file and search then.